requirements.txt
aiofiles~=24.1.0
autogen-agentchat~=0.4.9
autogen-ext~=0.4.9
fastapi~=0.115.11
openai~=1.66.3
openpyxl~=3.1.5
pandas~=2.2.3
python-dotenv~=1.0.1
pyyaml~=6.0.2
tiktoken~=0.9.0
uvicorn~=0.34.0
